\name{PREGS-package}
\alias{PREGS-package}
\alias{PREGS}
\docType{package}
\title{
\packageTitle{PREGS}
}
\description{
\packageDescription{PREGS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PREGS}
\packageIndices{PREGS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PREGS}

Maintainer: \packageMaintainer{PREGS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
